move is_url_requirement for use in parsing
